+++ index.html
@@ -228,6 +228,5 @@
   </template>
 
 </body>
-<script src="/js/functions.js"> </script>
-<script src="/js/main.js"> </script>
+<script src="/js/main.js" type="module"> </script>
 </html>
